refactor: TableGroup 생성 시 검증 로직 TableGroupValidator로 위임
